fixup! added immediate export to elastic on product update in administration
