+++ docs/upgrade/UPGRADE-unreleased.md
@@ -134,9 +134,13 @@ There you can find links to upgrade notes for other versions too.
                         type: "null"
         +               excluded_404s: false
         ```
-- if you have `Shopsys\FrameworkBundle\Model\Product\ProductFacade` extended in your `services.yml` and/or `services_test.yml`, add proper setter calls ([#1153](https://github.com/shopsys/shopsys/pull/1153))
+- if you have `Shopsys\FrameworkBundle\Model\Product\ProductFacade` or `Shopsys\FrameworkBundle\Model\Product\ProductVariantFacade` extended in your `services.yml` and/or `services_test.yml`, add proper setter calls ([#1153](https://github.com/shopsys/shopsys/pull/1153))
     ```diff
         Shopsys\FrameworkBundle\Model\Product\ProductFacade
+    +       calls:
+    +           - method: setProductSearchExportScheduler
+
+        Shopsys\FrameworkBundle\Model\Product\ProductVariantFacade
     +       calls:
     +           - method: setProductSearchExportScheduler
     ```
