+++ packages/framework/src/Model/Product/ProductFacade.php
@@ -2,6 +2,7 @@
 
 namespace Shopsys\FrameworkBundle\Model\Product;
 
+use BadMethodCallException;
 use Doctrine\ORM\EntityManagerInterface;
 use Shopsys\FrameworkBundle\Component\Domain\Domain;
 use Shopsys\FrameworkBundle\Component\Image\ImageFacade;
@@ -18,6 +19,7 @@ use Shopsys\FrameworkBundle\Model\Product\Pricing\ProductManualInputPriceFacade;
 use Shopsys\FrameworkBundle\Model\Product\Pricing\ProductPriceCalculation;
 use Shopsys\FrameworkBundle\Model\Product\Pricing\ProductPriceRecalculationScheduler;
 use Shopsys\FrameworkBundle\Model\Product\Pricing\ProductSellingPrice;
+use Shopsys\FrameworkBundle\Model\Product\Search\Export\ProductSearchExportScheduler;
 
 class ProductFacade
 {
@@ -132,6 +134,11 @@ class ProductFacade
     protected $productPriceCalculation;
 
     /**
+     * @var \Shopsys\FrameworkBundle\Model\Product\Search\Export\ProductSearchExportScheduler
+     */
+    protected $productSearchExportScheduler;
+
+    /**
      * @param \Doctrine\ORM\EntityManagerInterface $em
      * @param \Shopsys\FrameworkBundle\Model\Product\ProductRepository $productRepository
      * @param \Shopsys\FrameworkBundle\Model\Product\ProductVisibilityFacade $productVisibilityFacade
@@ -204,6 +211,20 @@ class ProductFacade
     }
 
     /**
+     * @param \Shopsys\FrameworkBundle\Model\Product\Search\Export\ProductSearchExportScheduler $productSearchExportScheduler
+     * @internal Will be replaced with constructor injection in the next major release
+     * @deprecated
+     */
+    public function setProductSearchExportScheduler(ProductSearchExportScheduler $productSearchExportScheduler): void
+    {
+        if ($this->productSearchExportScheduler !== null) {
+            throw new BadMethodCallException(sprintf('Method "%s" has been already called and cannot be called multiple times.', __METHOD__));
+        }
+
+        $this->productSearchExportScheduler = $productSearchExportScheduler;
+    }
+
+    /**
      * @param int $productId
      * @return \Shopsys\FrameworkBundle\Model\Product\Product
      */
@@ -225,6 +246,9 @@ class ProductFacade
         $this->setAdditionalDataAfterCreate($product, $productData);
 
         $this->pluginCrudExtensionFacade->saveAllData('product', $product->getId(), $productData->pluginData);
+
+        $this->validateInjectedDependencies();
+        $this->productSearchExportScheduler->scheduleProductIdForImmediateExport($product->getId());
 
         return $product;
     }
@@ -288,6 +312,13 @@ class ProductFacade
         $this->productVisibilityFacade->refreshProductsVisibilityForMarkedDelayed();
         $this->productPriceRecalculationScheduler->scheduleProductForImmediateRecalculation($product);
 
+        $productIdToExport = $product->getId();
+        if ($product->isVariant()) {
+            $productIdToExport = $product->getMainVariant()->getId();
+        }
+        $this->validateInjectedDependencies();
+        $this->productSearchExportScheduler->scheduleProductIdForImmediateExport($productIdToExport);
+
         return $product;
     }
 
@@ -296,14 +327,21 @@ class ProductFacade
      */
     public function delete($productId)
     {
+        $this->validateInjectedDependencies();
+
         $product = $this->productRepository->getById($productId);
         $productDeleteResult = $product->getProductDeleteResult();
         $productsForRecalculations = $productDeleteResult->getProductsForRecalculations();
+
+        $this->productSearchExportScheduler->scheduleProductIdForImmediateExport($productId);
+
         foreach ($productsForRecalculations as $productForRecalculations) {
             $this->productPriceRecalculationScheduler->scheduleProductForImmediateRecalculation($productForRecalculations);
             $productForRecalculations->markForVisibilityRecalculation();
             $this->productAvailabilityRecalculationScheduler->scheduleProductForImmediateRecalculation($productForRecalculations);
+            $this->productSearchExportScheduler->scheduleProductIdForImmediateExport($productForRecalculations->getId());
         }
+
         $this->em->remove($product);
         $this->em->flush();
 
@@ -441,5 +479,16 @@ class ProductFacade
     public function getOneByCatnumExcludeMainVariants($productCatnum)
     {
         return $this->productRepository->getOneByCatnumExcludeMainVariants($productCatnum);
+    }
+
+    /**
+     * @internal Will be removed in the next major release
+     * @deprecated
+     */
+    protected function validateInjectedDependencies(): void
+    {
+        if (!$this->productSearchExportScheduler instanceof ProductSearchExportScheduler) {
+            throw new BadMethodCallException(sprintf('Method "%s::setProductSearchExportScheduler()" has to be called in "services.yml" definition.', __CLASS__));
+        }
     }
 }
