+++ packages/framework/src/Resources/config/services_test.yml
@@ -11,10 +11,6 @@ services:
         arguments:
             - '%overwrite_domain_url%'
 
-    Shopsys\FrameworkBundle\Controller\Admin\ProductController:
-        calls:
-            - method: setProductSearchExportScheduler
-
     Shopsys\FrameworkBundle\Model\Administrator\Activity\AdministratorActivityFacade: ~
 
     Shopsys\FrameworkBundle\Model\Administrator\AdministratorRepository: ~
@@ -166,7 +162,9 @@ services:
     Shopsys\FrameworkBundle\Model\Product\ProductDataFactoryInterface:
         alias: Shopsys\FrameworkBundle\Model\Product\ProductDataFactory
 
-    Shopsys\FrameworkBundle\Model\Product\ProductFacade: ~
+    Shopsys\FrameworkBundle\Model\Product\ProductFacade:
+        calls:
+            - method: setProductSearchExportScheduler
 
     Shopsys\FrameworkBundle\Model\Product\ProductFactory: ~ # @deprecated
 
@@ -179,7 +177,9 @@ services:
 
     Shopsys\FrameworkBundle\Model\Product\ProductVisibilityRepository: ~
 
-    Shopsys\FrameworkBundle\Model\Product\ProductVariantFacade: ~
+    Shopsys\FrameworkBundle\Model\Product\ProductVariantFacade:
+        calls:
+            - method: setProductSearchExportScheduler
 
     Shopsys\FrameworkBundle\Model\Product\ProductOnCurrentDomainFacade: ~
